requirements.txt
protobuf
google-cloud-bigquery
pandas~=2.3.3
matplotlib~=3.10.8
seaborn~=0.13.2
db-dtypes
